Dev (#74)
* Migrate uv (#54)

* Move project to uv

* Move justfile to uv

* Move docker image to uv

* Update justfile

* Update workflow to use UV

* Sync version with pyproject.toml

* Zip files (#55)

* Move project to uv

* Move justfile to uv

* Move docker image to uv

* Update justfile

* Add functions for working with zip files

* pygrader now can work with zip archieves as input

* Version bump

* Move clone path for functional tests

* Bump version in uv.lock

* Add coverage via uv

* Add tests for unzip

* Fix unit tests for unzip

* Add functional test

* Web view split (#57)

* Update pyproject

* Remove web from this repo

* Update CHANGELOG

* Pylintrc part of config (#68)

* Add support for passing pylintrc as part of the config

* Update tests

* Update uv.lock

* Bump version to 1.7.0

* Add pylintrc example

* Remove old requirements.txt and HW2 leftovers

* Move sphinx stuff to uv

* Update uv.lock

* Add total score and max score to JSON, CSV, and PlainText result displays

* Update CHANGELOG. Version in constants is now taken from pyproject.toml

* Environment variables can now be passed from the config

* Update CHANGELOG and add config info in documentation

* Refactor environment variable handling across checks and tests

* Fix PR pipeline

* Fix failing unit tests

* mypy check requires venv

* pyproject.toml is a constant

* requirements check supports pyproject.toml

* Fix functional tests

* Update CI/CD pipelines to not use the predefined actions as they look ugly

+++ tests/test_functional.py
@@ -5,7 +5,9 @@ Module for functional tests of the grader.
 import os
 import shutil
 import unittest
+import zipfile
 
+from pathlib import Path
 from typing import Optional
 
 import grader.utils.constants as const
@@ -18,7 +20,7 @@ class BaseFunctionalTestWithGrader(unittest.TestCase):
     """
 
     repo_url = "https://github.com/fmipython/pygrader"
-    clone_path = "/tmp/pygrader"
+    clone_path = "/tmp/pygrader-cloned"
 
     def setUp(self) -> None:
         if os.path.exists(self.clone_path):
@@ -258,7 +260,11 @@ class TestFunctionalBadWeatherWithGrader(BaseFunctionalTestWithGrader):
         # Arrange
         command = build_command(project_path=self.clone_path)
 
-        os.remove(os.path.join(self.clone_path, "requirements.txt"))
+        if os.path.exists(os.path.join(self.clone_path, "requirements.txt")):
+            os.remove(os.path.join(self.clone_path, "requirements.txt"))
+
+        if os.path.exists(os.path.join(self.clone_path, "pyproject.toml")):
+            os.remove(os.path.join(self.clone_path, "pyproject.toml"))
 
         # Act
         run_result = run(command)
@@ -504,6 +510,33 @@ class TestRemoteTests(BaseFunctionalTestWithSampleProject):
         # Assert
         self.assertEqual(run_result.returncode, 0, run_result.stdout)
         self.assertTrue(is_score_correct(expected_score=13.5, target_check="tests", grader_output=run_result.stdout))
+
+
+class TestZipFileOnSampleProject(BaseFunctionalTestWithSampleProject):
+    def test_01_zip_archive_passed(self) -> None:
+        """
+        Verify that when passing a zip version of the project, it is graded.
+        """
+        # Arrange
+        folder_path = Path(self.clone_path)
+        zip_path = folder_path / "project.zip"
+
+        with zipfile.ZipFile(zip_path, "w", zipfile.ZIP_DEFLATED) as zip_file:
+            for entry in folder_path.rglob("*"):
+                if entry.is_file():
+                    zip_file.write(entry, entry.relative_to(folder_path))
+
+        command = build_command(project_path=str(zip_path), config_file="only_pylint.json")
+
+        # Act
+        run_result = run(command)
+
+        # Assert
+        self.assertEqual(run_result.returncode, 0, run_result.stdout)
+        self.assertTrue(
+            is_score_correct(expected_score=1, target_check="pylint", grader_output=run_result.stdout),
+            "Pylint check did not have the expected score of 1",
+        )
 
 
 def build_command(
